% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expit}
\alias{expit}
\title{inverse logit function}
\usage{
expit(x)
}
\arguments{
\item{x}{A real number}
}
\value{
expit(x) = \code{1/(1+exp(-x))}
}
\description{
inverse logit function
}
\examples{
expit(0.5)
logit(seq(from=-10,to=+10,length=9))
}
