% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{fit.model.em}
\alias{fit.model.em}
\title{Maximum likelihood estimation for simple row clustering
without covariates - maximisation using the EM algorithm}
\usage{
fit.model.em(
  ymat,
  R,
  itermin = 10,
  itermax = 100,
  aeps = 1e-06,
  use.gradient = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{ymat}{Matrix of binary data}

\item{R}{Number of clusters to fit}

\item{itermax}{Maximum number of EM iterations}

\item{aeps}{Absolute tolerance for convergence of parameters}

\item{use.gradient}{(logical) Use the gradient of the log likelihood?}

\item{verbose}{(logical) Print out evaluations of the log likelihood
and its derivatives?}
}
\value{
Fitted model object
}
\description{
Maximum likelihood estimation for simple row clustering
without covariates - maximisation using the EM algorithm
}
