% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{fit.model}
\alias{fit.model}
\title{Maximum likelihood estimation for simple row clustering
without covariates - direct maximisation of the incomplete
data log likelihood}
\usage{
fit.model(ymat, R, use.gradient = TRUE, verbose = FALSE)
}
\arguments{
\item{ymat}{Matrix of binary data}

\item{R}{Number of clusters to fit}

\item{use.gradient}{(logical) Use the gradient of the log likelihood?}

\item{verbose}{(logical) Print out evaluations of the log likelihood
and its derivatives?}
}
\value{
Fitted model object
}
\description{
Maximum likelihood estimation for simple row clustering
without covariates - direct maximisation of the incomplete
data log likelihood
}
