% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{grad.llikef.complete.rbclust}
\alias{grad.llikef.complete.rbclust}
\title{Gradient of the complete data log L}
\usage{
grad.llikef.complete.rbclust(model, n, long.df, xmat.df, zmat)
}
\description{
Gradient of the complete data log L
}
