% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{grad.llikef}
\alias{grad.llikef}
\title{Gradient of the log likelihood function - simple row clustering
without covariates}
\usage{
grad.llikef(ymat, model)
}
\description{
Gradient of the log likelihood function - simple row clustering
without covariates
}
