% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_llikef_complete_rbclust_c}
\alias{grad_llikef_complete_rbclust_c}
\title{C++ version of the gradient of the complete data log L}
\usage{
grad_llikef_complete_rbclust_c(model, n, long_df, xmat_df, zmat)
}
\description{
C++ version of the gradient of the complete data log L
}
