% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_llikef_rbclust_c}
\alias{grad_llikef_rbclust_c}
\title{C++ version of the gradient of log L for
row clustering for binary data}
\usage{
grad_llikef_rbclust_c(model, n, long_df, xmat_df)
}
\description{
C++ version of the gradient of log L for
row clustering for binary data
}
