% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{llikef}
\alias{llikef}
\title{Log likelihood function - simple row clustering
without covariates}
\usage{
llikef(ymat, model)
}
\description{
Log likelihood function - simple row clustering
without covariates
}
