% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llikef_complete_rbclust_c}
\alias{llikef_complete_rbclust_c}
\title{C++ version of the complete data log L for
row clustering for binary data}
\usage{
llikef_complete_rbclust_c(model, n, long_df, xmat_df, zmat)
}
\description{
C++ version of the complete data log L for
row clustering for binary data
}
