% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llikef_rbclust_c}
\alias{llikef_rbclust_c}
\title{C++ version of log L of row clustering for binary data}
\usage{
llikef_rbclust_c(model, n, long_df, xmat_df)
}
\description{
C++ version of log L of row clustering for binary data
}
