% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{A number between 0 and 1}
}
\value{
logit(p) = \code{log(p/(1-p))}
}
\description{
logit function
}
\examples{
logit(0.5)
logit(seq(from=0.1,to=0.9,length=8))
}
