% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{make.model}
\alias{make.model}
\title{Interpret the parameter vector for the model with
simple row clustering without covariates}
\usage{
make.model(parvec, model)
}
\description{
Interpret the parameter vector for the model with
simple row clustering without covariates
}
