% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{make.parvec}
\alias{make.parvec}
\title{Load the parameters of the model with simple row clustering
without covariates into a vector}
\usage{
make.parvec(model)
}
\description{
Load the parameters of the model with simple row clustering
without covariates into a vector
}
