% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{make.xmat.rbclustmodel}
\alias{make.xmat.rbclustmodel}
\title{Design matrix of the row clustered model with covariates}
\usage{
make.xmat.rbclustmodel(
  formula,
  R,
  data,
  xr.df = NULL,
  xc.df = NULL,
  responsename = "Y",
  factorname1 = "Row",
  factorname2 = "Col",
  rowclustername = "RowClust"
)
}
\description{
Design matrix of the row clustered model with covariates
}
