% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{predictmodel.rbclust}
\alias{predictmodel.rbclust}
\title{Prediction for the row clustered model with covariates}
\usage{
predictmodel.rbclust(
  model,
  nnew,
  newxr.df = NULL,
  responsename = "Y",
  factorname1 = "Row",
  factorname2 = "Col",
  rowclustername = "RowClust"
)
}
\description{
Prediction for the row clustered model with covariates
}
