% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{simdata}
\alias{simdata}
\title{Simulate a dataset}
\usage{
simdata(n, model)
}
\arguments{
\item{n}{number of rows}

\item{model}{model specification: see details}
}
\value{
a list object with two entries
   \code{kvec} a vector of length \code{n} containing the row
   group label for each row
   \code{ymat} an \code{n}x\code{m} matrix of binary outcomes
   generated in the simulation
}
\description{
Generate n independent rows of a matrix specified by the
object \code{model}
}
\details{
The model specification object \code{model} contains
   \code{R} the number of cluster components
   \code{pi} the mixing probabilities -
     a vector of length R, adding to 1, with non-negative entries
   \code{theta} the probabilities of success in each row group
     - a vector of length R of probabilities (values between 0 and 1)
   \code{m} the number of columns
}
